apps/api/requirements.txt
Flask==3.1.0
flask-cors==5.0.0
flask-limiter==3.8.0
python-dotenv==1.0.1
firebase-admin==6.5.0
gunicorn==23.0.0
pydantic==2.10.5
PyJWT==2.9.0
bcrypt==4.2.1
email-validator==2.2.0
